requirements.txt
fastapi==0.112.2
httpx==0.27.0
pydantic==2.8.2
tortoise-orm==0.22.1
uvicorn==0.32.1
